Reorganize the project package structure
