+++ src/main/java/com/fmachinus/practice/entity/dto/CustomerDto.java
@@ -1,4 +1,4 @@
-package com.fmachinus.practice.customer;
+package com.fmachinus.practice.entity.dto;
 
 public class CustomerDto {
